+++ pages/api/historical.ts
@@ -1,12 +1,6 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+import { getHistoricalData } from "@coin-view/api";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-const marketMap = {
-  USD: "binance-us",
-  PLN: "binance",
-};
-
-const periods = 3600;
 
 export default async function handler(
   req: NextApiRequest,
@@ -20,23 +14,10 @@ export default async function handler(
     return;
   }
 
-  const date = new Date();
-  date.setDate(date.getDate() - 1);
-
-  const yesterdayTimestamp = Math.floor(date.getTime() / 1000);
-
-  const market = marketMap[currency];
-
-  const promises = symbols.split(",").map(async (symbol) => {
-    const response = await fetch(
-      `https://api.cryptowat.ch/markets/${market}/${symbol}${currency}/ohlc?periods=${periods}&after=${yesterdayTimestamp}`
-    );
-
-    const data = await response.json();
-    return [symbol, data.result[periods]];
+  const data = await getHistoricalData({
+    currency,
+    symbols: symbols.split(","),
   });
 
-  const data = await Promise.all(promises);
-
-  res.status(200).json(Object.fromEntries(data));
+  res.status(200).json(data);
 }
