+++ src/api/index.ts
@@ -1,2 +1,3 @@
 export * from "./getCoinList";
 export * from "./getCoinsMetadata";
+export * from "./getHistoricalData";
