+++ src/api/mockHistoricalData.json
@@ -1,0 +1,104 @@
+{
+  "3600": [
+    [
+      1669658400, 16302.2, 16378.19, 16185.3, 16199.69, 922.330603,
+      15009493.28850413
+    ],
+    [
+      1669662000, 16200.2, 16227.35, 16137.79, 16154.44, 405.902605,
+      6563411.61007705
+    ],
+    [
+      1669665600, 16154.4, 16244.76, 16142.28, 16215.12, 379.32638,
+      6141846.57134441
+    ],
+    [
+      1669669200, 16218.23, 16263.73, 16203.2, 16251.6, 257.397917,
+      4177501.2705917
+    ],
+    [
+      1669672800, 16250, 16255.36, 16181.99, 16194.01, 275.085161,
+      4459693.7023624
+    ],
+    [
+      1669676400, 16194.57, 16228.66, 16190.17, 16212.35, 101.934353,
+      1652431.80880175
+    ],
+    [
+      1669680000, 16216.57, 16239.63, 16188.97, 16206.23, 191.520499,
+      3105157.92647118
+    ],
+    [
+      1669683600, 16204.95, 16219.13, 16094.54, 16170.04, 514.66309,
+      8312040.62060283
+    ],
+    [
+      1669687200, 16170.91, 16296, 16165.31, 16242.75, 270.156223,
+      4384563.62119556
+    ],
+    [
+      1669690800, 16243.67, 16298.19, 16238.61, 16255.06, 240.01759,
+      3903597.80963142
+    ],
+    [
+      1669694400, 16256.71, 16297.06, 16250.01, 16278.68, 134.39373,
+      2187044.74049362
+    ],
+    [
+      1669698000, 16279.65, 16477.38, 16271.5, 16421.07, 730.441963,
+      11967016.43769067
+    ],
+    [
+      1669701600, 16423.58, 16526.86, 16422.57, 16474.07, 446.860004,
+      7365053.26707237
+    ],
+    [
+      1669705200, 16473.7, 16506.26, 16438.31, 16450, 107.531322,
+      1770156.83076371
+    ],
+    [
+      1669708800, 16455.14, 16495.06, 16446.11, 16455, 103.35407,
+      1702266.96640222
+    ],
+    [
+      1669712400, 16454.69, 16538.91, 16420.66, 16490.83, 262.568707,
+      4328887.50628213
+    ],
+    [
+      1669716000, 16488.34, 16516.85, 16469.57, 16507.05, 78.882975,
+      1300429.32669172
+    ],
+    [
+      1669719600, 16509.2, 16525.1, 16478.27, 16497.98, 134.04145,
+      2211449.76740255
+    ],
+    [
+      1669723200, 16498.01, 16534.15, 16480.87, 16491.29, 102.029429,
+      1683479.65435613
+    ],
+    [
+      1669726800, 16491.29, 16502.1, 16462.48, 16479.99, 128.81425,
+      2122717.66714818
+    ],
+    [
+      1669730400, 16483.62, 16483.62, 16375.29, 16407.58, 238.85538,
+      3923501.57180131
+    ],
+    [
+      1669734000, 16404.75, 16454.5, 16325.2, 16353.85, 462.324991,
+      7583357.20528085
+    ],
+    [
+      1669737600, 16353.62, 16448.1, 16337.04, 16385.51, 431.587657,
+      7072271.08089281
+    ],
+    [
+      1669741200, 16387.95, 16421.12, 16337.04, 16403.56, 473.215633,
+      7751472.57774902
+    ],
+    [
+      1669744800, 16405.98, 16422.27, 16352.47, 16384.09, 185.091775,
+      3033365.91596216
+    ]
+  ]
+}
